Improved DataEntryApp and started adding integration testing with WebApplicationFactory

+++ DataEntryApp/Program.cs
@@ -4,59 +4,67 @@ using System.Net.Http.Json;
 using var client = new HttpClient();
 
 client.BaseAddress = new Uri("https://localhost:7069/");
+
 try
 {
     HttpResponseMessage response = await client.GetAsync("teams");
     response.EnsureSuccessStatusCode();
     string jsonResponse = await response.Content.ReadAsStringAsync();
     var teams = JsonSerializer.Deserialize<List<TeamData>>(jsonResponse, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
-    RunRequests(teams);
+    bool running = true;
+    while (running)
+    {
+        running = RunRequests(teams); // RunRequests() returns true for every request except Stop, so the user can continue entering requests until they'd like to stop.
+    }
 }
 catch (HttpRequestException e)
 {
     Console.WriteLine($"Error: {e.Message}");
 }
 
-void RunRequests(List<TeamData> teams){
-    Console.WriteLine("What would you like to do? Type PrintAll, PrintByID, PrintByConference, PrintByState, DeleteByID, AddTeam, UpdateByID");
-    string request = Console.ReadLine();
+bool RunRequests(List<TeamData> teams){
+    Console.WriteLine("What would you like to do? Type PrintAll, PrintByID, PrintByConference, PrintByState, DeleteByID, AddTeam, UpdateByID, or Stop to stop.");
+    string request = Console.ReadLine().ToUpper();
     switch (request)
     {
-        case "PrintAll":
+        case "PRINTALL":
             PrintAll(teams);
-            break;
-        case "PrintByID":
+            return true;
+        case "PRINTBYID":
             Console.WriteLine("Enter the team's ID: ");
             var id = Console.ReadLine();
             PrintByID(teams, id);
-            break;
-        case "PrintByConference":
+            return true;
+        case "PRINTBYCONFERENCE":
             Console.WriteLine("Enter the conference name: ");
             var conference = Console.ReadLine();
             PrintByConference(teams, conference);
-            break;
-        case "PrintByState":
+            return true;
+        case "PRINTBYSTATE":
             Console.WriteLine("Enter the state name: ");
             var state = Console.ReadLine();
             PrintByState(teams, state);
-            break;
-        case "DeleteByID":
+            return true;
+        case "DELETEBYID":
             Console.WriteLine("Enter the team's ID: ");
             var delId = Console.ReadLine();
             DeleteByID(teams, delId);
-            break;
-        case "AddTeam":
+            return true;
+        case "ADDTEAM":
             var newID = teams.Count() + 1;
             AddTeam(teams, newID);
-            break;
-        case "UpdateByID":
+            return true;
+        case "UPDATEBYID":
             Console.WriteLine("Enter the team's ID: ");
             var updateID = Console.ReadLine();
             UpdateTeam(teams, updateID);
-            break;
+            return true;
+        case "STOP":
+            return false;
         default:
             Console.WriteLine("Please enter a valid request.");
-            break;
+            RunRequests(teams);
+            return true;
     }
 }
 
@@ -68,6 +76,10 @@ void PrintAll(List<TeamData> teams)
             $"Data for {team.TeamName} ({team.TeamId}):\nLocation: {team.City}, {team.State}\nConference: {team.Conference}"
         );
     }
+    if (teams.Count == 0)
+    {
+        Console.WriteLine("No teams found.");
+    }
 }
 
 void PrintByID(List<TeamData> teams, string id)
@@ -75,10 +87,21 @@ void PrintByID(List<TeamData> teams, string id)
     if(int.TryParse(id, out int IdInt))
     {
         var found = teams.FindAll(team => team.TeamId == IdInt);
+        foreach (TeamData foundTeam in found)
+        {
+            Console.WriteLine(
+                $"Data for {foundTeam.TeamName} ({foundTeam.TeamId}):\nLocation: {foundTeam.City}, {foundTeam.State}\nConference: {foundTeam.Conference}"
+                );
+        }
+        if (found.Count == 0)
+        {
+        Console.WriteLine("No teams found with that ID.");
+        }
     }
     else
     {
         Console.WriteLine("Please enter a valid number.");
+        PrintByID(teams, id);
     }
 }
 
@@ -91,6 +114,10 @@ void PrintByConference(List<TeamData> teams, string conference)
                 $"Data for {foundTeam.TeamName} ({foundTeam.TeamId}):\nLocation: {foundTeam.City}, {foundTeam.State}\nConference: {foundTeam.Conference}"
                 );
     }
+    if (found.Count == 0)
+    {
+        Console.WriteLine("No teams found in that conference.");
+    }
 }
 
 void PrintByState(List<TeamData> teams, string state)
@@ -102,6 +129,10 @@ void PrintByState(List<TeamData> teams, string state)
                 $"Data for {foundTeam.TeamName} ({foundTeam.TeamId}):\nLocation: {foundTeam.City}, {foundTeam.State}\nConference: {foundTeam.Conference}"
                 );
     }
+    if (found.Count == 0)
+    {
+        Console.WriteLine("No teams found in that state.");
+    }
 }
 
 async void DeleteByID(List<TeamData> teams, string id)
@@ -110,13 +141,21 @@ async void DeleteByID(List<TeamData> teams, string id)
     if(int.TryParse(id, out int IdInt))
     {
         var found = teams.Find(team => team.TeamId == IdInt);
-        teams.Remove(found);
-        HttpResponseMessage postResponse = await client.DeleteAsync($"teams/{IdInt}");
-        postResponse.EnsureSuccessStatusCode();
+        if (found == null)
+        {
+            Console.WriteLine("No team with that ID was found.");
+        }
+        else
+        {
+            teams.Remove(found);
+            HttpResponseMessage postResponse = await client.DeleteAsync($"teams/{IdInt}");
+            postResponse.EnsureSuccessStatusCode();
+        }
     }
     else
     {
         Console.WriteLine("Please enter a valid number.");
+        DeleteByID(teams, id);
     }
 
     
@@ -153,50 +192,57 @@ async void UpdateTeam(List<TeamData> teams, string id)
     if(int.TryParse(id, out int IdInt))
     {
         Console.WriteLine("What would you like to update? Enter Name, City, State, or Conference");
-        string request = Console.ReadLine();
-        switch (request){
-            case "Name":
-                var foundForName = teams.FindAll(team => team.TeamId == IdInt);
-                Console.WriteLine("Enter the new team name: ");
-                var newName = Console.ReadLine();
-                foundForName[0].TeamName = newName;
-                HttpResponseMessage namePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForName[0]);
-                namePutResponse.EnsureSuccessStatusCode();
-                break;
-            case "City":
-                var foundForCity = teams.FindAll(team => team.TeamId == IdInt);
-                Console.WriteLine("Enter the new team city: ");
-                var newCity = Console.ReadLine();
-                foundForCity[0].City = newCity;
+        bool validField = false;
+        while (!validField)
+        {
+            string request = Console.ReadLine();
+            switch (request){
+                case "Name":
+                    var foundForName = teams.FindAll(team => team.TeamId == IdInt);
+                    Console.WriteLine("Enter the new team name: ");
+                    var newName = Console.ReadLine();
+                    foundForName[0].TeamName = newName;
+                    HttpResponseMessage namePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForName[0]);
+                    namePutResponse.EnsureSuccessStatusCode();
+                    validField = true;
+                    break;
+                case "City":
+                    var foundForCity = teams.FindAll(team => team.TeamId == IdInt);
+                    Console.WriteLine("Enter the new team city: ");
+                    var newCity = Console.ReadLine();
+                    foundForCity[0].City = newCity;
 
-                HttpResponseMessage cityPutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForCity[0]);
-                cityPutResponse.EnsureSuccessStatusCode();
-                
-                break;
-            case "State":
-                var foundForState = teams.FindAll(team => team.TeamId == IdInt);
-                Console.WriteLine("Enter the new team state: ");
-                var newState = Console.ReadLine();
-                foundForState[0].State = newState;
+                    HttpResponseMessage cityPutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForCity[0]);
+                    cityPutResponse.EnsureSuccessStatusCode();
+                    validField = true;
+                    break;
+                case "State":
+                    var foundForState = teams.FindAll(team => team.TeamId == IdInt);
+                    Console.WriteLine("Enter the new team state: ");
+                    var newState = Console.ReadLine();
+                    foundForState[0].State = newState;
 
-                HttpResponseMessage statePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForState[0]);
-                statePutResponse.EnsureSuccessStatusCode();
+                    HttpResponseMessage statePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForState[0]);
+                    statePutResponse.EnsureSuccessStatusCode();
+                    validField = true;
+                    break;
+                case "Conference":
+                    var foundForConference = teams.FindAll(team => team.TeamId == IdInt);
+                    Console.WriteLine("Enter the new team conference: ");
+                    var newConference = Console.ReadLine();
+                    foundForConference[0].Conference = newConference;
 
-                break;
-            case "Conference":
-                var foundForConference = teams.FindAll(team => team.TeamId == IdInt);
-                Console.WriteLine("Enter the new team conference: ");
-                var newConference = Console.ReadLine();
-                foundForConference[0].Conference = newConference;
-
-                HttpResponseMessage conferencePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForConference[0]);
-                conferencePutResponse.EnsureSuccessStatusCode();
-
-                break;
-            default:
-                Console.WriteLine("Please enter a valid field.");
-                break;
+                    HttpResponseMessage conferencePutResponse = await client.PutAsJsonAsync($"teams/{id}", foundForConference[0]);
+                    conferencePutResponse.EnsureSuccessStatusCode();
+                    validField = true;
+                    break;
+                default:
+                    Console.WriteLine("Please enter a valid field.");
+                    validField = false;
+                    break;
+            }
         }
+        
 
     }
     else
